Move Enemy Intent Collector to AI Module
