+++ src/enemy_ai.rs
@@ -1,29 +1,19 @@
 //! Enemy fighter AI
 
 use bevy::prelude::*;
-use iyes_loopless::prelude::*;
 use rand::{prelude::SliceRandom, Rng};
 
 use crate::{
+    animation::Facing,
+    commands::CustomCommands,
     consts,
     enemy::{Enemy, TripPointX},
-    fighter_state::{FighterStateCollectSystems, Idling},
+    fighter_state::{
+        Attacking, Idling, Moving, StateTransition, StateTransitionIntents, TransitionCmds,
+    },
     player::Player,
-    GameState,
+    Stats,
 };
-
-pub struct EnemyAiPlugin;
-
-impl Plugin for EnemyAiPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_system_to_stage(
-            CoreStage::PreUpdate,
-            set_target_near_player
-                .run_in_state(GameState::InGame)
-                .before(FighterStateCollectSystems),
-        );
-    }
-}
 
 /// A place that an enemy fighter is going to move to
 #[derive(Component)]
@@ -33,7 +23,10 @@ pub struct EnemyTarget {
 }
 
 // For enemys without current target, pick a new spot near the player as target
-fn set_target_near_player(
+///
+/// This is added to the [`crate::fighter_state::FighterStateCollectSystems`] to collect figher
+/// actions for enemies.
+pub fn set_target_near_player(
     mut commands: Commands,
     mut enemies_query: Query<
         (Entity, &mut TripPointX),
@@ -65,6 +58,58 @@ fn set_target_near_player(
                     });
                 }
             }
+        }
+    }
+}
+
+/// Controls enemy AI fighters
+///
+/// This is added to the [`crate::fighter_state::FighterStateCollectSystems`] to collect figher
+/// actions for enemies.
+pub fn emit_enemy_intents(
+    mut query: Query<
+        (
+            Entity,
+            &Transform,
+            &Stats,
+            &EnemyTarget,
+            &mut Facing,
+            &mut StateTransitionIntents,
+        ),
+        // All enemies that are either moving or idling
+        (With<Enemy>, Or<(With<Idling>, With<Moving>)>),
+    >,
+    mut transition_commands: CustomCommands<TransitionCmds>,
+    time: Res<Time>,
+) {
+    let mut commands = transition_commands.commands();
+
+    for (entity, transform, stats, target, mut facing, mut intents) in &mut query {
+        let position = transform.translation.truncate();
+        let velocity =
+            (target.position - position).normalize() * stats.movement_speed * time.delta_seconds();
+
+        if velocity.x < 0.0 {
+            *facing = Facing::Left;
+        } else {
+            *facing = Facing::Right;
+        }
+
+        // If we're close to our target
+        if position.distance(target.position) <= 100. {
+            // Remove the target
+            commands.entity(entity).remove::<EnemyTarget>();
+
+            // And attack!
+            intents.push_back(StateTransition::new(
+                Attacking::default(),
+                Attacking::PRIORITY,
+            ));
+
+        // If we aren't near our target yet
+        } else {
+            // Move towards our target
+            intents.push_back(StateTransition::new(Moving { velocity }, Moving::PRIORITY));
         }
     }
 }
